# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Cartlabs API
  version: 0.2.0
  description: Public REST contract for Cartlabs web and API applications.
servers:
  - url: /api/v1
...
security:
  - BearerAuth: []
paths:
...
  /auth/demo-login:
    post:
      operationId: demoLogin
      summary: Authenticate as demo role
      tags: [Identity]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/DemoLoginRequest"}
      responses:
        "200":
          description: Authenticated as deterministic demo account
          headers:
            Set-Cookie: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/TooManyRequests"}
        "503": {$ref: "#/components/responses/ServiceUnavailableProblem"}
...
  /admin/products/{productId}/moderation:
    patch:
      operationId: moderateProduct
      summary: Moderate product
      tags: [Admin]
      parameters:
        - {$ref: "#/components/parameters/ProductId"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ModerationInput"}
      responses:
        "200":
          description: Moderated product
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProductDetail"}
        "404": {$ref: "#/components/responses/NotFound"}
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    RefreshCookie:
      type: apiKey
      in: cookie
      name: __Host-cartlabs_refresh
      description: Production Secure cookie with Path=/. Local HTTP uses cartlabs_refresh scoped to authentication routes.
  parameters:
    ProductId: {name: productId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
...
    Forbidden: {$ref: "#/components/responses/Problem403"}
    NotFound: {$ref: "#/components/responses/Problem404"}
...
    TooManyRequests: {$ref: "#/components/responses/Problem429"}
...
    ServiceUnavailableProblem: {$ref: "#/components/responses/Problem503JSON"}
...
    Problem403: {description: Insufficient permission, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    Problem404: {description: Resource not found, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
...
    Problem429: {description: Authentication attempts exceeded, headers: {Retry-After: {schema: {type: integer}}}, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
...
    Problem503JSON: {description: Authentication limiter unavailable, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        requestId: {type: string}
    Role: {type: string, enum: [buyer, seller, admin]}
    ModerationStatus: {type: string, enum: [pending, approved, rejected]}
...
    DemoLoginRequest:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role: {$ref: "#/components/schemas/Role"}
    Session:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken: {type: string}
        tokenType: {type: string, enum: [Bearer]}
        expiresIn: {type: integer, minimum: 1}
        user: {$ref: "#/components/schemas/User"}
    User:
      type: object
      required: [id, email, displayName, role]
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        displayName: {type: string}
        role: {$ref: "#/components/schemas/Role"}
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        slug: {type: string}
...
    VariantInput:
      type: object
      additionalProperties: false
      required: [sku, name, attributes, priceMinor, currency, stock]
      properties:
        sku: {type: string, minLength: 2, maxLength: 64}
        name: {type: string, minLength: 1, maxLength: 120}
        attributes: {type: object, additionalProperties: {type: string}}
        priceMinor: {type: integer, format: int64, minimum: 0}
        currency: {type: string, enum: [IDR]}
        stock: {type: integer, minimum: 0}
    Variant:
      allOf:
        - {$ref: "#/components/schemas/VariantInput"}
        - type: object
          required: [id, active]
          properties:
            id: {type: string, format: uuid}
            active: {type: boolean}
    ImageInput:
      type: object
      additionalProperties: false
      required: [url, altText, position]
      properties:
        url: {type: string, format: uri-reference, pattern: '^(https?://|/[^/])'}
        altText: {type: string, maxLength: 160}
        position: {type: integer, minimum: 0}
    ProductImage:
      allOf:
        - {$ref: "#/components/schemas/ImageInput"}
        - type: object
          required: [id]
          properties:
            id: {type: string, format: uuid}
...
    ProductDetail:
      type: object
      required: [id, storeId, category, name, slug, description, status, moderationStatus, moderationNote, variants, images, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        storeId: {type: string, format: uuid}
        storeName: {type: string}
        category: {$ref: "#/components/schemas/Category"}
        name: {type: string}
        slug: {type: string}
        description: {type: string}
        status: {type: string, enum: [draft, published, archived]}
        moderationStatus: {$ref: "#/components/schemas/ModerationStatus"}
        moderationNote: {type: string}
        variants: {type: array, items: {$ref: "#/components/schemas/Variant"}}
        images: {type: array, items: {$ref: "#/components/schemas/ProductImage"}}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
...
    ModerationInput:
      type: object
      additionalProperties: false
      required: [status, note]
      properties:
        status: {type: string, enum: [approved, rejected]}
        note: {type: string, maxLength: 500}